# ci/test: speed up tests