--- a/.github/workflows/hive.yml
+++ b/.github/workflows/hive.yml
@@ -29,7 +29,7 @@ jobs:
         with:
           context: .
           tags: ghcr.io/paradigmxyz/reth:latest
-          build-args: BUILD_PROFILE=hivetests
+          build-args: BUILD_PROFILE=hivetests,FEATURES=asm-keccak
           outputs: type=docker,dest=./artifacts/reth_image.tar
           cache-from: type=gha
           cache-to: type=gha,mode=max
@@ -198,7 +198,7 @@ jobs:
           echo "Simulator failed, creating issue"
           # Check if issue already exists
           # get all issues with the label C-hivetest, loop over each page and check if the issue already exists
-          
+
           existing_issues=$(gh api /repos/paradigmxyz/reth/issues -H "Accept: application/vnd.github+json" -H "X-GitHub-Api-Version: 2022-11-28" -F "labels=C-hivetest" --method GET | jq '.[].title')
           if [[ $existing_issues == *"Hive Test Failure: ${{ matrix.sim }}"* ]]; then
             echo "Issue already exists"
@@ -211,7 +211,7 @@ jobs:
             /repos/${{ github.repository }}/issues \
             -f title='Hive Test Failure: ${{ matrix.sim }}' \
             -f body="!!!!!!! This is an automated issue created by the hive test failure !!!!!!!<br /><br />The hive test for ${{ matrix.sim }} failed. Please investigate and fix the issue.<br /><br />[Link to the failed run](https://github.com/paradigmxyz/reth/actions/runs/${{ github.run_id }})" \
-            -f "labels[]=C-hivetest" 
+            -f "labels[]=C-hivetest"
       - name: Print simulator output
         if: ${{ failure() }}
         run: |