--- a/.github/workflows/integration.yml
+++ b/.github/workflows/integration.yml
@@ -40,10 +40,10 @@ jobs:
       - name: Run tests
         run: |
           cargo nextest run \
-            --locked --features "${{ matrix.network }}" \
+            --locked --features "asm-keccak ${{ matrix.network }}" \
             --workspace --exclude examples --exclude ef-tests \
             --partition hash:${{ matrix.partition }}/2 \
-            -E 'kind(test)'
+            -E "kind(test)"
 
   sync:
     name: sync / 100k blocks
@@ -63,7 +63,7 @@ jobs:
           cache-on-failure: true
       - name: Run sync
         run: |
-          cargo run --release --features jemalloc,min-error-logs --bin reth \
+          cargo run --release --features asm-keccak,jemalloc,min-error-logs --bin reth \
             -- node \
             --debug.tip 0x91c90676cab257a59cd956d7cb0bceb9b1a71d79755c23c7277a0697ccfaf8c4 \
             --debug.max-block 100000 \