--- a/crates/storage/db/src/implementation/mdbx/tx.rs
+++ b/crates/storage/db/src/implementation/mdbx/tx.rs
@@ -415,12 +415,13 @@ mod tests {
         // Give the `TxnManager` some time to time out the transaction.
         sleep(MAX_DURATION + Duration::from_millis(100));
 
+        // Transaction has not timed out.
         assert_eq!(
-            tx.get::<tables::Transactions>(0).err(),
-            Some(DatabaseError::Open(reth_libmdbx::Error::NotFound.into()))
-        ); // Transaction is not timeout-ed
+            tx.get::<tables::Transactions>(0),
+            Err(DatabaseError::Open(reth_libmdbx::Error::NotFound.into()))
+        );
+        // Backtrace is not recorded.
         assert!(!tx.metrics_handler.unwrap().backtrace_recorded.load(Ordering::Relaxed));
-        // Backtrace is not recorded
     }
 
     #[test]
@@ -437,11 +438,12 @@ mod tests {
         // Give the `TxnManager` some time to time out the transaction.
         sleep(MAX_DURATION + Duration::from_millis(100));
 
+        // Transaction has timed out.
         assert_eq!(
-            tx.get::<tables::Transactions>(0).err(),
-            Some(DatabaseError::Open(reth_libmdbx::Error::ReadTransactionTimeout.into()))
-        ); // Transaction is timeout-ed
+            tx.get::<tables::Transactions>(0),
+            Err(DatabaseError::Open(reth_libmdbx::Error::ReadTransactionTimeout.into()))
+        );
+        // Backtrace is recorded.
         assert!(tx.metrics_handler.unwrap().backtrace_recorded.load(Ordering::Relaxed));
-        // Backtrace is recorded
     }
 }