--- a/testing/ef-tests/Cargo.toml
+++ b/testing/ef-tests/Cargo.toml
@@ -13,6 +13,7 @@ workspace = true
 
 [features]
 ef-tests = []
+asm-keccak = ["reth-primitives/asm-keccak"]
 
 [dependencies]
 reth-primitives.workspace = true