We use proptest and keccak256 a lot, so make them way faster for a very small, insignificant cost to compile time, which is cached anyway in CI.

Some slow tests on my machine and CI
- `SealedBlockTests::proptest`: 20s to 2s, CI [60s](https://github.com/paradigmxyz/reth/actions/runs/8165780763/job/22323474765#step:6:671), to [3.2s](https://github.com/paradigmxyz/reth/actions/runs/8166425981/job/22325173127?pr=6987#step:6:464)
- `timestamped_forks`: 12.5s to 1.5s, CI [42.5s](https://github.com/paradigmxyz/reth/actions/runs/8165780763/job/22323474765#step:6:669), to [4.7s](https://github.com/paradigmxyz/reth/actions/runs/8166425981/job/22325173127?pr=6987#step:6:605)
- `fuzz_state_root_incremental`: 11.1s to 2.15s, CI [25s](https://github.com/paradigmxyz/reth/actions/runs/8165780763/job/22323474765#step:6:672), to [4.2s](https://github.com/paradigmxyz/reth/actions/runs/8166425981/job/22325173127?pr=6987#step:6:746)
- `prune_receipts_by_logs`: 3.5s to 0.5s


~~There are 2 tests that are hard coded to sleep 60s which is not ideal (`long_read_transaction_safety_disabled`), added in https://github.com/paradigmxyz/reth/pull/6181 by @shekhirin. I don't know enough about that part of the code base to reduce their running time.~~ https://github.com/paradigmxyz/reth/pull/6994

`stage_checkpoint_pruned` and `execute_intermediate_commit` are flaky, the first run it times out, then runs fine.

Remaining slow tests are quite a few networking tests that also look hard coded to 5 / 10 seconds, like [`txn_manager_read_transactions_duration_unbounded`](https://github.com/paradigmxyz/reth/actions/runs/8166425981/job/22325173127?pr=6987#step:6:371), [`test_backoff_on_no_response`](https://github.com/paradigmxyz/reth/actions/runs/8166425981/job/22325173127?pr=6987#step:6:388), cc @emhane @mattsse if there's anything we can do to make them faster, or if there's even a point in doing so.

---

I've removed partitions for integration tests since the total running time is <30s with a 5m compile time, there is no point in splitting them.